syntex analyzer parsing writen and partialy tested

+++ src/Production.cpp
@@ -20,7 +20,7 @@ string Production::toString() {
     }
 
     string ans = LHS->name+" -> ";
-    for(int i = 0; i < RHS->size(); i++) {
+    for(int i = 0; i < (int)RHS->size(); i++) {
         ans += RHS->at(i)->name+" ";
     }
